@@ -10,15 +10,15 @@
     "animate.css-react": "^1.1.0",
     "aos": "^2.3.4",
     "netlify-cli": "^2.12.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.1.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.10.0",
     "react-scripts": "^2.1.3"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
+    "start": "react-scripts --openssl-legacy-provider start",
+    "build": "react-scripts --openssl-legacy-provider build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
