@@ -1,5 +1,5 @@
 import React from "react"
-import ReactDOM from "react-dom"
+import {createRoot} from "react-dom/client"
 import "./index.css"
 import App from "./App"
 
@@ -11,9 +11,10 @@ window.onscroll = () => {
   else navbar.className = ""
 }
 
-ReactDOM.render(
+const root = createRoot(document.getElementById("root"))
+
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 )
